+++ lib/minfra/cli.rb
@@ -51,6 +51,7 @@ module Minfra
     end
 
     def self.exec(argv)
+     
       init(argv) unless init?
       cli.install
       cli.run